+++ .dir-locals.el
@@ -2,9 +2,4 @@
   .
   ((cider-clojure-cli-aliases
     .
-    "dev")))
- (clojure-mode
-  .
-  ((eval
-    .
-    (add-to-list 'cider-jack-in-nrepl-middlewares "scicloj.clay.v1.nrepl/middleware")))))
+    "dev"))))